Moved FakeWebsocket to a helper
